tree: move try_resolve_file_conflict() to tree_merge()
